internal: move api.d.ts to base-core
