+++ generators/base-core/support/write-files.spec.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 import { describe, esmocha, expect, it } from 'esmocha';
-import type { EditFileCallback } from '../../base/api.js';
+import type { EditFileCallback } from '../../base-core/api.js';
 import { joinCallbacks } from './write-files.js';
 
 describe('generator - base - support - writeFiles', () => {
