+++ generators/java/support/files.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import type { WriteFileBlock } from '../../base/api.js';
+import type { WriteFileBlock } from '../../base-core/api.js';
 import type CoreGenerator from '../../base-core/generator.js';
 import { SERVER_MAIN_RES_DIR, SERVER_MAIN_SRC_DIR, SERVER_TEST_RES_DIR, SERVER_TEST_SRC_DIR } from '../../generator-constants.js';
 import type { ApplicationType } from '../../../lib/types/application/application.js';
