+++ generators/spring-cloud-stream/generators/pulsar/files.ts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { WriteFileSection } from '../../../base/api.js';
+import type { WriteFileSection } from '../../../base-core/api.js';
 import { GRADLE_BUILD_SRC_MAIN_DIR, SERVER_TEST_SRC_DIR } from '../../../generator-constants.js';
 import { moveToJavaPackageTestDir } from '../../../java/support/index.js';
